requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
Pillow
requests
